+++ CHANGELOG.md
@@ -16,7 +16,7 @@ without native/platform specific bindings and SDKs. See [this ticket for more de
 
 API Changes:
 
-- Removed IHasDistribution ([#2660](https://github.com/getsentry/sentry-dotnet/pull/2660))
+- Adding `Distribution` to `IEventLike` ([#2660](https://github.com/getsentry/sentry-dotnet/pull/2660))
 
 ### Features