+++ CHANGELOG.md
@@ -20,6 +20,7 @@ API Changes:
 - ISpanContext was removed. Use ITraceContext instead. ([#2668](https://github.com/getsentry/sentry-dotnet/pull/2668))
 - Removed IHasTransactionNameSource. Use ITransactionContext instead. ([#2654](https://github.com/getsentry/sentry-dotnet/pull/2654))
 - Adding `Distribution` to `IEventLike` ([#2660](https://github.com/getsentry/sentry-dotnet/pull/2660))
+- Removed IHasTags and IHasExtra. Use ITagsAndData instead ([#2675](https://github.com/getsentry/sentry-dotnet/pull/2675))
 
 ### Features