+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Dependencies
+
+- Bump CLI from v2.20.7 to v2.21.0 ([#2645](https://github.com/getsentry/sentry-dotnet/pull/2645))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2210)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.20.7...2.21.0)
+
 ## 3.40.0-beta.0
 
 ### Breaking changes